# book-network-ui/src/app/services/models/book-response.ts
/* tslint:disable */
/* eslint-disable */
export interface BookResponse {
  archived?: boolean;
  authorName?: string;
  bookCover?: string;
  isbn?: string;
  ownerFirstName?: string;
  ownerLastName?: string;
  rate?: number;
  shareable?: boolean;
  synopsis?: string;
  title?: string;
  uuid?: string;
}