# book-network-ui/src/app/services/models/borrowed-book-response.ts
/* tslint:disable */
/* eslint-disable */
export interface BorrowedBookResponse {
  bookAuthorName?: string;
  bookAverageRating?: number;
  bookIsbn?: string;
  bookTitle?: string;
  bookUuid?: string;
  returnApproved?: boolean;
  returned?: boolean;
}